k-space density uses multiprocesses
